Add swift_check_params config
In some clouds (pre Queens) the url for Swift endpoint checks needs to
be different.  This config param allows a custom url to be used, to
accommodate the wide variety of services which might provide a 'swift'
endpoint.

This change also corrects the urls used for placement to be compatible
across Pike to Queens, and the Ceilometer url which is only used up to
Pike.