Add config-changed
Add a flag to ensure that when config-changed runs, the nrpe checks are
reconfigured.  Prior behaviour was that nrpe checks do not get
re-written resulting in changes to the keystone catalog getting missed.
While this does not actively search for changes to the keystone catalog,
at least a config-changed hook will pick it up.