Allow endpoint URLs without an explicit port
Endpoint URLs listed in Keystone catalog may not provide an explicit
port. By default, ports 80 or 443 would be used, and this change fixes
it. On the other hand, glance-simplestreams-sync (image-stream service)
may or may not use Swift (use_swift=True), so it is not known if the
configured endpoint will be reachable. This change also skips checks
against the image-stream service (note that if swift is used, a
different endpoint will exist with the same URL, and it will be
checked).

Closes-Bug: 1835952