drop trusty support
this charm has following deps in layer.yaml:

  - python3-keystoneauth1
  - python3-keystoneclient
  - python3-openstackclient
  - python3-neutronclient

But none of these python3-* packages are available on trusty.

Also, it's also using snap layer to install fcbtest, which will break on trusty as well.

In fact, we already lost trusty support.
To avoid consusion, this patch remove trusty in metadata.yaml.

LP: #1877923

Signed-off-by: Joe Guo <joe.guo@canonical.com>